token-balances: improved interface for querying token balances
